move GNUstep.h into common.h
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@29727 72102866-910b-0410-8b05-ffd578937521

+++ Source/Additions/GCObject.m
@@ -450,7 +450,7 @@ static pthread_mutex_t *allocationLock = NULL;
 /**
  * Returns the receivers reference count.
  */
-- (unsigned int) retainCount
+- (NSUInteger) retainCount
 {
   return gc.flags.refCount;
 }
